chore(wallet-mobile): Refactor connected dApps queries
